ci/github-script: split getting commit details for PR into its own file
